+++ samples/src/main/java/jp/shiguredo/sora/sample/config/Config.kt
@@ -1,9 +1,0 @@
-package jp.shiguredo.sora.sample.config
-
-class Config {
-
-    companion object {
-        const val SIGNALING_ENDPOINT = "wss://sora.example.com/signaling"
-    }
-}
-