+++ samples/src/main/java/jp/shiguredo/sora/sample/ui/VideoChatRoomActivity.kt
@@ -11,8 +11,8 @@ import android.view.Gravity
 import android.view.View
 import android.view.WindowManager
 import android.widget.*
+import jp.shiguredo.sora.sample.BuildConfig
 import jp.shiguredo.sora.sample.R
-import jp.shiguredo.sora.sample.config.Config
 import jp.shiguredo.sora.sample.facade.SoraVideoChannel
 import jp.shiguredo.sora.sample.option.SoraStreamType
 import jp.shiguredo.sora.sample.ui.util.RendererLayoutCalculator
@@ -205,7 +205,7 @@ class VideoChatRoomActivity : AppCompatActivity() {
 
         channel = SoraVideoChannel(
                 context           = this,
-                signalingEndpoint = Config.SIGNALING_ENDPOINT,
+                signalingEndpoint = BuildConfig.SIGNALING_ENDPOINT,
                 channelId         = channelName,
                 signalingMetadata = "",
                 videoWidth        = videoWidth,